Docs.AsHtml: Fix doc folding symmetry
Folded source includes the summary in the details, because we currently
don't have a good way to separate a signature from a definition or even
a shortened type (as seen in the Fuzzy Finder) from its definition.

This commit changes the non-source version of Folded to also include its
summary in the details part, such that the 2 Folded implementations at
least are symmetric (even though they are both semantically incorrect in
the eyes of html and the <details> element).